+++ file_scraper/__init__.py
@@ -1,2 +1,2 @@
 """file-scraper module."""
-__version__ = '0.44'
+__version__ = '0.45'